1.1 patch. Battles now are a little more random with ~1.5x army will win but lower can lose. Much better realm text placement. New win and defeat sounds. Fixed a lot of warnings.

+++ DegreeProject/DegreeProject/Source/Game/AI/Managers/AIRelationshipManager.cpp
@@ -76,7 +76,7 @@ void AIRelationshipManager::makeAlliance(AIData& data)
 float AIRelationshipManager::marriageDecision(AIData& data, CharacterID potentialSpouse)
 {
 	MarriageConsideration marriage;
-	float opinionWeight = CharacterManager::get().getCharacterOpinion(data.m_OwnerID, potentialSpouse);
+	float opinionWeight = (float)CharacterManager::get().getCharacterOpinion(data.m_OwnerID, potentialSpouse);
 	return marriage.evaluate(data.m_OwnerID, potentialSpouse) + (opinionWeight * 0.1f);
 }
